Update3
.Defined the distanceFell function
. Called the distanceFell function in MenuOptions.cpp if choice = 3
. Changed the input variables in the getUserInput function to a double type

+++ Useful Functions.h
@@ -7,7 +7,7 @@ inline void clearErrorState() {
 }
 
 inline double getUserInput() {
-    int input;
+    double input;
     std::cout << "Enter a value:" << std::endl;
     while (true) {
         if (std::cin >> input) {
